requirements.txt
asgiref==3.7.2
awscli==1.32.21
boto3==1.34.21
botocore==1.34.21
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.4
cryptography==41.0.7
Django==5.0.1
django-extensions==3.2.3
django-jazzmin==2.6.0
docutils==0.16
idna==3.6
jmespath==1.0.1
Naked==0.1.32
pyasn1==0.5.1
pycparser==2.21
pycryptodome==3.20.0
pydotplus==2.0.2
pygraphviz==1.12
pyparsing==3.1.1
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.10.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
